Added some cuda stuff

+++ include/data/typedefs.h
@@ -7,15 +7,15 @@
 
 namespace Aperture {
 
-typedef double Scalar;
+typedef float Scalar;
 
 /// A simple way to track staggering of field components. Note that when
 /// initializing it, one can use Stagger_t("011"). The rightmost digit is for x
 /// direction, while leftmost digit for z
 typedef std::bitset<3> Stagger_t;
 
-typedef double Mom_t;
-typedef double Pos_t;
+typedef float Mom_t;
+typedef float Pos_t;
 
 typedef std::size_t Index_t;
